+++ vyper/semantics/analysis/module.py
@@ -4,6 +4,7 @@ from typing import Optional, Union
 
 import vyper.builtins.interfaces
 from vyper import ast as vy_ast
+from vyper.ast.pre_typecheck import prefold
 from vyper.evm.opcodes import version_check
 from vyper.exceptions import (
     CallViolation,
@@ -28,7 +29,7 @@ from vyper.semantics.data_locations import DataLocation
 from vyper.semantics.namespace import Namespace, get_namespace
 from vyper.semantics.types import EnumT, EventT, InterfaceT, StructT
 from vyper.semantics.types.function import ContractFunctionT
-from vyper.semantics.types.utils import prefold, type_from_annotation
+from vyper.semantics.types.utils import type_from_annotation
 from vyper.typing import InterfaceDict
