+++ vyper/semantics/analysis/utils.py
@@ -2,6 +2,7 @@ import itertools
 from typing import Callable, List
 
 from vyper import ast as vy_ast
+from vyper.ast.pre_typecheck import prefold
 from vyper.exceptions import (
     CompilerPanic,
     InvalidLiteral,
@@ -24,7 +25,6 @@ from vyper.semantics.types.base import TYPE_T, VyperType
 from vyper.semantics.types.bytestrings import BytesT, StringT
 from vyper.semantics.types.primitives import AddressT, BoolT, BytesM_T, IntegerT
 from vyper.semantics.types.subscriptable import DArrayT, SArrayT, TupleT
-from vyper.semantics.types.utils import prefold
 from vyper.utils import checksum_encode, int_to_fourbytes
