+++ vyper/semantics/types/subscriptable.py
@@ -3,12 +3,13 @@ from typing import Any, Dict, Optional, Tuple, Union
 
 from vyper import ast as vy_ast
 from vyper.abi_types import ABI_DynamicArray, ABI_StaticArray, ABI_Tuple, ABIType
+from vyper.ast.pre_typecheck import prefold
 from vyper.exceptions import ArrayIndexException, InvalidType, StructureException
 from vyper.semantics.data_locations import DataLocation
 from vyper.semantics.types.base import VyperType
 from vyper.semantics.types.primitives import IntegerT
 from vyper.semantics.types.shortcuts import UINT256_T
-from vyper.semantics.types.utils import get_index_value, prefold, type_from_annotation
+from vyper.semantics.types.utils import get_index_value, type_from_annotation
 
 
 class _SubscriptableT(VyperType):
